refactor: merge api
