+++ taipei-day-trip/api/attractions.py
@@ -5,7 +5,8 @@ import db
 
 atts_blueprints = Blueprint( 'atts', __name__ )
 
-@atts_blueprints.route('/api/attractions', methods=['GET'])
+# 取得景點資料列表
+@atts_blueprints.route('/attractions', methods=['GET'])
 def Attractions():
     try:
         page = request.args.get('page')
@@ -58,6 +59,61 @@ def Attractions():
                 "error": True,
                 "message": "伺服器內部錯誤"
             }
+        res = make_response(jsonify(data),500)
+    finally:
+        return res 
+    
+# 根據景點編號取得景點資料
+@atts_blueprints.route('/attraction/<int:attractionId>', methods=['GET'])
+def AttractionId(attractionId):
+    try:
+        result = db.select_id(attractionId)
+        if result[0] == None:
+            raise ValueError("無此景點編號，請重新嘗試")
+        data = {
+            "data": {
+                "id": result[0],
+                "name": result[1],
+                "category": result[2],
+                "description": result[3],
+                "address": result[4],
+                "transport": result[5],
+                "mrt": result[6],
+                "lat": result[7],
+                "lng": result[8],
+                "images": result[9].split(',')
+            }
+        }
+        res = make_response(jsonify(data),200)
+    except ValueError as msg:
+        data = {
+            "error": True,
+            "message": str(msg)
+        }
+        res = make_response(jsonify(data),400)
+    except:
+        data = {
+            "error": True,
+            "message": "伺服器內部錯誤"
+        }
+        res = make_response(jsonify(data),500)
+    finally:
+        return res 
+    
+# 取得景點分類名稱列表
+@atts_blueprints.route('/categories', methods=['GET'])
+def Categories():
+    try:
+        result = db.select_cat()
+        data = {
+            "data": result[0][0].split(',')
+        }
+        res = make_response(jsonify(data),200)
+    except:
+        data = {
+            "error": True,
+            "message": "伺服器內部錯誤"
+        }
         res = make_response(jsonify(data),500)
     finally:
         return res
